Helpers for converting model specifications in R to Julia equivalents

Helpers for converting model specifications in R to Julia equivalents

Usage

jl_formula(x)

jl_contrasts(df, cols = NULL, show_code = FALSE)

jl_data(df)

jl_family(family = c("gaussian", "binomial", "poisson"))

Arguments

x

A string or formula object

df

A data frame

cols

A subset of columns to make contrast specifiations for

show_code

Whether to print corresponding Julia code as a side-effect

family

The distributional family as string or <family> object

Examples

if (nzchar(Sys.which("julia"))) {
  jlme_setup(restart = TRUE)

  # (Setup) Model data in R
  x <- mtcars
  x$cyl_helm <- factor(x$cyl)
  contrasts(x$cyl_helm) <- contr.helmert(3)
  colnames(contrasts(x$cyl_helm)) <- c("4vs6", "4&6vs8")

  # Formula conversion with
  julia_formula <- jl_formula(mpg ~ am * cyl_helm)
  julia_formula

  # Data frame conversion
  julia_data <- jl_data(x)
  julia_data

  # Contrasts construction (`show_code = TRUE` pretty prints the Julia code)
  julia_contrasts <- jl_contrasts(x, show_code = TRUE)
  julia_contrasts

  # Family conversion
  julia_family <- jl_family("binomial")
  julia_family

  stop_julia()
}
#> Dict(
#>     :cyl_helm => HypothesisCoding(
#>         [
#>             -1/2  1/2   0
#>             -1/6 -1/6 1/3
#>         ];
#>         levels = ["4", "6", "8"],
#>         labels = ["4vs6", "4&6vs8"],
#>     ),
#> )