Set up Julia connection for jlme

Set up Julia connection for jlme

Usage

check_julia_ok()

stop_julia()

jlme_status()

jlme_setup(
  ...,
  add = NULL,
  restart = FALSE,
  threads = NULL,
  verbose = interactive()
)

Arguments

...

Unused

add

A character vector of additional Julia packages to add and load.

restart

Whether to run stop_julia() first, before attempting setup

threads

Number of threads to start Julia with. Defaults to 1

verbose

Whether to alert setup progress. Defaults to interactive()

Value

Invisibly returns TRUE on success

Examples

# \donttest{
# Check whether Julia installation meets requirements
check_julia_ok()
#> [1] TRUE

# Connect to a Julia runtime for use with `{jlme}`
jlme_setup()
#> Starting Julia (v1.10.5) ...
#> Successfully set up Julia connection. (19s)

# Show information about the Julia runtime
jlme_status()
#> jlme 0.4.1 
#> R version 4.4.1 (2024-06-14 ucrt) 
#> Julia Version 1.10.5
#> Commit 6f3fdf7b36 (2024-08-27 14:19 UTC)
#> Build Info:
#>   Official https://julialang.org/ release
#> Platform Info:
#>   OS: Windows (x86_64-w64-mingw32)
#>   CPU: 8 × 11th Gen Intel(R) Core(TM) i7-1165G7 @ 2.80GHz
#>   WORD_SIZE: 64
#>   LIBM: libopenlibm
#>   LLVM: libLLVM-15.0.7 (ORCJIT, tigerlake)
#> Threads: 1 default, 0 interactive, 1 GC (on 8 virtual cores)
#> Status `C:\Users\jchoe\AppData\Local\Temp\jl_3cLpwU\Project.toml`
#>   [38e38edf] GLM v1.9.0
#>   [ff71e718] MixedModels v4.31.0
#>   [3eaba693] StatsModels v0.7.4

# Stop Julia runtime
stop_julia()
# }