Profile the likelihood surface of Julia mixed effects models

Profile the likelihood surface of Julia mixed effects models

Usage

profilelikelihood(x, ...)

Arguments

x

A Julia MixedModel of class jlme

...

Not implemented

Value

MixedModels.profile() output as object of class jlmeprof

Examples

# \donttest{
jlme_setup(restart = TRUE)
#> Starting Julia (v1.10.5) ...
#> Successfully set up Julia connection.

jmod <- jlmer(Reaction ~ Days + (Days | Subject), lme4::sleepstudy)
tidy(jmod)
#> # A tibble: 6 × 7
#>   effect   group    term                 estimate std.error statistic    p.value
#>   <chr>    <chr>    <chr>                   <dbl>     <dbl>     <dbl>      <dbl>
#> 1 fixed    NA       (Intercept)          251.          6.63     37.9   2.02e-314
#> 2 fixed    NA       Days                  10.5         1.50      6.97  3.22e- 12
#> 3 ran_pars Subject  sd__(Intercept)       23.8        NA        NA    NA        
#> 4 ran_pars Subject  cor__(Intercept).Da…   0.0813     NA        NA    NA        
#> 5 ran_pars Subject  sd__Days               5.72       NA        NA    NA        
#> 6 ran_pars Residual sd__Observation       25.6        NA        NA    NA        

prof <- profilelikelihood(jmod)
prof
#> <Julia object of type MixedModelProfile{Float64}>
#> MixedModelProfile -- Table with 11 columns and 176 rows:
#>       p  ζ          β1       β2       σ        σ1       σ2       ρ1           ⋯
#>     ┌──────────────────────────────────────────────────────────────────────────
#>  1  │ σ  -4.365     251.405  10.4673  20.1933  25.5128  5.97319  -0.0159232   ⋯
#>  2  │ σ  -3.77902   251.405  10.4673  20.8002  25.3434  5.94786  -0.00711109  ⋯
#>  3  │ σ  -3.20526   251.405  10.4673  21.4255  25.1627  5.92133  0.00263598   ⋯
#>  4  │ σ  -2.64336   251.405  10.4673  22.0695  24.9702  5.89214  0.0129078    ⋯
#>  5  │ σ  -2.09298   251.405  10.4673  22.7328  24.7636  5.86167  0.0241968    ⋯
#>  6  │ σ  -1.55378   251.405  10.4673  23.4162  24.5426  5.82871  0.0366992    ⋯
#>  7  │ σ  -1.02542   251.405  10.4673  24.12    24.3063  5.79389  0.0501628    ⋯
#>  8  │ σ  -0.507597  251.405  10.4673  24.845   24.0525  5.75669  0.0649969    ⋯
#>  9  │ σ  0.0        251.405  10.4673  25.5918  23.7805  5.71683  0.0813321    ⋯
#>  10 │ σ  0.497684   251.405  10.4673  26.3611  23.4885  5.6743   0.0993046    ⋯
#>  11 │ σ  0.985728   251.405  10.4673  27.1534  23.1743  5.62882  0.119204     ⋯
#>  12 │ σ  1.46442    251.405  10.4673  27.9696  22.836   5.5802   0.141262     ⋯
#>  13 │ σ  1.93402    251.405  10.4673  28.8104  22.4727  5.52809  0.165869     ⋯
#>  14 │ σ  2.39481    251.405  10.4673  29.6763  22.0799  5.47226  0.193415     ⋯
#>  15 │ σ  2.84704    251.405  10.4673  30.5684  21.6557  5.41243  0.224387     ⋯
#>  16 │ σ  3.29094    251.405  10.4673  31.4872  21.196   5.34818  0.259442     ⋯
#>  17 │ σ  3.72677    251.405  10.4673  32.4337  20.6956  5.27882  0.299408     ⋯
#>  ⋮  │ ⋮      ⋮         ⋮        ⋮        ⋮        ⋮        ⋮          ⋮       ⋱

tidy(prof)
#> # A tibble: 5 × 6
#>   effect   group    term            estimate  lower  upper
#>   <chr>    <chr>    <chr>              <dbl>  <dbl>  <dbl>
#> 1 fixed    NA       (Intercept)       251.   238.   265.  
#> 2 fixed    NA       Days               10.5    7.36  13.6 
#> 3 ran_pars Subject  sd__(Intercept)    23.8   14.4   37.7 
#> 4 ran_pars Subject  sd__Days            5.72   0      8.75
#> 5 ran_pars Residual sd__Observation    25.6   22.9   28.9 

stop_julia()
# }