Helpers for converting model specifications in R to Julia equivalents

Helpers for converting model specifications in R to Julia equivalents

Usage

is_jl(x, type)

jl_put(x)

jl_get(x)

jl(expr, ..., .R = FALSE, .passthrough = FALSE)

jl_dict(...)

jl_formula(formula)

jl_contrasts(df, cols = NULL, show_code = FALSE)

jl_data(df)

jl_family(family = c("gaussian", "binomial", "poisson"))

Arguments

x

An object

type

Type of Julia object to additional test for

expr

A string of Julia code

...

Interpolated elements. In the case of jl():.

If all named, elements are introduced as Julia variables in the expr

If all unnamed, elements are interpolated into expr via sprintf()

.R

Whether to simplify and return as R object, if possible.

.passthrough

Whether to return expr as-is if it's already a Julia object. Mostly for internal use.

formula

A string or formula object

df

A data frame

cols

A subset of columns to make contrast specifiations for

show_code

Whether to print corresponding Julia code as a side-effect

family

The distributional family as string or <family> object

Value

A Julia object of type <JuliaProxy>

Examples

# \donttest{
jlme_setup(restart = TRUE)
#> [dev] Using prior setup: `is_setup()` is TRUE in `load_all()`
#> [dev] Using prior setup: `is_setup()` is TRUE in `load_all()`
#> Starting Julia (v1.10.5) ...
#> Successfully set up Julia connection.

# (general) Use `jl()` to evaluate arbitrary Julia expressions from string
jl("1 .+ [1,3]")
#> <Julia object of type Vector{Int64}>
#> 2-element Vector{Int64}:
#>  2
#>  4

# `jl()` takes elements in `...` that you can reference in the expression
jl("1 .+ a", a = c(1L, 3L)) # Named arguments are introduced as variables
#> <Julia object of type Vector{Int64}>
#> 2-element Vector{Int64}:
#>  2
#>  4
jl("1 .+ %s", "[1,2]") # Unnamed arguments are interpolated via `sprintf()`
#> <Julia object of type Vector{Int64}>
#> 2-element Vector{Int64}:
#>  2
#>  3

# `jl_dict()` constructs a Dictionary data type
jl_dict(age = 20:25, sex = c("M", "F"))
#> <Julia object of type Dict{Symbol, Any}>
#> Dict{Symbol, Any} with 2 entries:
#>   :age => [20, 21, 22, 23, 24, 25]
#>   :sex => ["M", "F"]
jl_dict(list(a = 1, b = 2)) # Splats when a single list is supplied
#> <Julia object of type Dict{Symbol, Any}>
#> Dict{Symbol, Any} with 2 entries:
#>   :a => [1.0]
#>   :b => [2.0]

# Use `is_jl()` to test if object is a Julia (`<JuliaProxy>`) object
is_jl(jl("1"))
#> [1] TRUE

# Use `jl_put()` and `jl_get()` to transfer data between R and Julia
jl_put(1L)
#> <Julia object of type Int64>
#> 1
identical(jl_get(jl_put(1L)), 1L)
#> [1] TRUE

# (modelling) set up model data in R
x <- mtcars
x$cyl_helm <- factor(x$cyl)
contrasts(x$cyl_helm) <- contr.helmert(3)
colnames(contrasts(x$cyl_helm)) <- c("4vs6", "4&6vs8")

# Formula conversion with
julia_formula <- jl_formula(mpg ~ am * cyl_helm)
julia_formula
#> <Julia object of type FormulaTerm{Term, Tuple{Term, Term, InteractionTerm{Tuple{Term, Term}}}}>
#> FormulaTerm
#> Response:
#>   mpg(unknown)
#> Predictors:
#>   am(unknown)
#>   cyl_helm(unknown)
#>   am(unknown) & cyl_helm(unknown)

# Data frame conversion
julia_data <- jl_data(x)
julia_data
#> <Julia object of type Main.RConnector.RDataFrame>
#> Main.RConnector.RDataFrame with 32 rows, 12 columns, and an unknown schema.

# Contrasts construction (`show_code = TRUE` pretty prints the Julia code)
julia_contrasts <- jl_contrasts(x, show_code = TRUE)
#> Dict(
#>     :cyl_helm => HypothesisCoding(
#>         [
#>             -1/2  1/2   0
#>             -1/6 -1/6 1/3
#>         ];
#>         levels = ["4", "6", "8"],
#>         labels = ["4vs6", "4&6vs8"],
#>     ),
#> )
julia_contrasts
#> <Julia object of type Dict{Symbol, HypothesisCoding{Matrix{Float64}, Matrix{Float64}}}>
#> Dict{Symbol, HypothesisCoding{Matrix{Float64}, Matrix{Float64}}} with 1 entry:
#>   :cyl_helm => HypothesisCoding{Matrix{Float64}, Matrix{Float64}}([-0.5 0.5 0.0…

# Family conversion
julia_family <- jl_family("binomial")
julia_family
#> <Julia object of type Bernoulli{Float64}>
#> Bernoulli{Float64}(p=0.5)

stop_julia()
# }